+++ molgraph/applications/proteomics/__init__.py
@@ -1,0 +1,11 @@
+from molgraph.applications.proteomics.peptide_models import PeptideGNN
+from molgraph.applications.proteomics.peptide_models import PeptideEmbedding
+from molgraph.applications.proteomics.peptide_saliency import PeptideSaliency
+from molgraph.applications.proteomics.peptide_encoders import PeptideGraphEncoder
+from molgraph.applications.proteomics.peptide_encoders import Peptide
+
+
+PepGNN = PeptideGNN
+PepSaliency = PeptideSaliency
+PepGraphEncoder = PeptideGraphEncoder
+PepEmbedding = PepEmbed = PeptideEmbedding